Adding labels and descriptions for scenes

+++ photons_interactor/commander/commands.py
@@ -1,8 +1,8 @@
 from photons_interactor.commander.errors import NoSuchCommand, NoSuchScene
+from photons_interactor.database.database import Scene, SceneInfo
 from photons_interactor.commander import default_fields as df
 from photons_interactor.commander.decorator import command
 from photons_interactor.commander import helpers as chp
-from photons_interactor.database.database import Scene
 
 from photons_app.errors import FoundNoDevices
 from photons_app import helpers as hp
@@ -171,37 +171,50 @@ class SetCommand(Command):
         return await chp.run(script, fltr, self.finder, timeout=self.timeout)
 
 @command(name="scene_info")
-class SceneInfo(Command):
+class SceneInfoCommand(Command):
     """
     Retrieve information about scenes in the database
     """
     db_queue = df.db_queue_field
 
-    uuids = dictobj.NullableField(sb.listof(sb.string_spec())
-        , help = "Only get information for scene with these uuids"
+    uuid = dictobj.NullableField(sb.listof(sb.string_spec())
+        , help = "Only get information for scene with these uuid"
         )
 
-    no_details = dictobj.Field(sb.boolean, default=False
-        , help = "Just return the uuids and labels"
+    only_meta = dictobj.Field(sb.boolean, default=False
+        , help = "Only return meta info about the scenes"
         )
 
     async def execute(self):
         def get(db):
-            info = defaultdict(list)
+            info = defaultdict(lambda: {"meta": {}, "scene": []})
+
             fs = []
-            if self.uuids:
-                fs.append(Scene.uuid.in_(self.uuids))
+            ifs = []
+            if self.uuid:
+                fs.append(Scene.uuid.in_(self.uuid))
+                ifs.append(SceneInfo.uuid.in_(self.uuid))
+
+            for sinfo in db.query(SceneInfo).filter(*ifs):
+                info[sinfo.uuid]["meta"] = sinfo.as_dict(ignore=["uuid"])
+
             for scene in db.query(Scene).filter(*fs):
-                dct = scene.as_dict()
-                del dct["uuid"]
-                info[scene.uuid].append(dct)
-            if self.no_details:
-                return sorted(info)
+                # Make sure there is an entry if no SceneInfo for this scene
+                info[scene.uuid]
+
+                if not self.only_meta:
+                    dct = scene.as_dict(ignore=["uuid"])
+                    info[scene.uuid]["scene"].append(dct)
+
+            if self.only_meta:
+                for _, data in info.items():
+                    del data["scene"]
+
             return dict(info)
         return await self.db_queue.request(get)
 
 @command(name="scene_change")
-class SceneChange(Command):
+class SceneChangeCommand(Command):
     """
     Set all the options for a scene
     """
@@ -211,25 +224,42 @@ class SceneChange(Command):
         , help = "The uuid of the scene to change, if None we create a new scene"
         )
 
-    scene = dictobj.Field(sb.listof(Scene.DelayedSpec(storing=True))
+    label = dictobj.NullableField(sb.string_spec
+        , help = "The label to give this scene"
+        )
+
+    description = dictobj.NullableField(sb.string_spec
+        , help = "The description to give this scene"
+        )
+
+    scene = dictobj.NullableField(sb.listof(Scene.DelayedSpec(storing=True))
         , help = "The options for the scene"
         )
 
     async def execute(self):
         def make(db):
             scene_uuid = self.uuid or str(uuid.uuid4())
-            for thing in db.queries.get_scenes(uuid=scene_uuid):
-                db.delete(thing)
 
-            for part in self.scene:
-                made = db.queries.create_scene(**part(scene_uuid).as_dict())
-                db.add(made)
+            if self.scene is not None:
+                for thing in db.queries.get_scenes(uuid=scene_uuid):
+                    db.delete(thing)
+
+                for part in self.scene:
+                    made = db.queries.create_scene(**part(scene_uuid).as_dict())
+                    db.add(made)
+
+            info, _ = db.queries.get_or_create_scene_info(uuid=scene_uuid)
+            if self.label is not None:
+                info.label = self.label
+            if self.description is not None:
+                info.description = self.description
+            db.add(info)
 
             return scene_uuid
         return await self.db_queue.request(make)
 
 @command(name="scene_delete")
-class SceneDelete(Command):
+class SceneDeleteCommand(Command):
     """
     Delete a scene
     """
@@ -248,7 +278,7 @@ class SceneDelete(Command):
         return await self.db_queue.request(delete)
 
 @command(name="scene_apply")
-class SceneApply(Command):
+class SceneApplyCommand(Command):
     """
     Apply a scene
     """
@@ -351,7 +381,7 @@ class SceneApply(Command):
         return clone
 
 @command(name="scene_capture")
-class SceneCapture(Command):
+class SceneCaptureCommand(Command):
     """
     Capture a scene
     """
@@ -364,6 +394,14 @@ class SceneCapture(Command):
 
     uuid = dictobj.NullableField(sb.string_spec
         , help = "The uuid of the scene to change, if None we create a new scene"
+        )
+
+    label = dictobj.NullableField(sb.string_spec
+        , help = "The label to give this scene"
+        )
+
+    description = dictobj.NullableField(sb.string_spec
+        , help = "The description to give this scene"
         )
 
     just_return = dictobj.Field(sb.boolean, default=False
@@ -420,5 +458,7 @@ class SceneCapture(Command):
         args = {
               "uuid": self.uuid
             , "scene": scene
+            , "label": self.label
+            , "description": self.description
             }
         return await self.commander.execute({"command": "scene_change", "args": args})
